requirements.txt
aiohttp>=3.8.0
pillow>=9.5.0
numpy>=1.24.0
Pillow>=10.0.0 